# Optimize performance and allocations of Scanner