This optimizes the performance of `Scanner` by avoiding allocations and manually unmarshalling the scan results into correctly typed values.

I also added a benchmark to measure the improvements.

Benchmark:

```
name       old time/op    new time/op    delta
Scanner-8    42.0ms ± 1%    32.2ms ± 1%  -23.32%  (p=0.008 n=5+5)

name       old alloc/op   new alloc/op   delta
Scanner-8    2.70MB ± 0%    0.55MB ± 0%  -79.65%  (p=0.008 n=5+5)

name       old allocs/op  new allocs/op  delta
Scanner-8     65.8k ± 0%     12.0k ± 0%  -81.83%  (p=0.008 n=5+5)
```