@nussjustin 2019-03-06
@mediocregopher Guess who needs to scan "a few" (*cough* million *cough*) keys and ran into some allocation problems.