@mediocregopher 2019-03-06
> Guess who needs to scan "a few" (cough million cough) keys and ran into some allocation problems.

Lmao, I'll bet I know who. I've never asked, but what are you using redis/radix for anyway?

Left one question about the `cur` field.