@nussjustin 2019-03-06
> I've never asked, but what are you using redis/radix for anyway?
I got more into Redis (and SQLite) in the last few months and started using it more and more, both at home and at work.

Mostly for smaller stuff, but I also wrote and maintain a new website for a customer with Redis as main data store and am currently working on migrating some parts of existing systems to Redis (especially now with Redis streams). Unfortunately most of the things at work are/must be written in PHP and not Go, but some smaller things are written in Go (e.g. when I need to work with much data).